Mg/Graph autoencoder
