+++ deeplay/components/gnn/mpn/get_edge_features.py
@@ -4,13 +4,12 @@ from .cla import CombineLayerActivation
 class GetEdgeFeatures(CombineLayerActivation):
     """"""
 
-    def get_forward_args(self, x):      # maybe use Tranform instead, and just take the first two ouputs
-        """Get the arguments for the ... module.
-        An MPN ... module takes the following arguments:
-        - node features of sender nodes (x[A[0]])
-        - node features of receiver nodes (x[A[1]])
+    def get_forward_args(self, x):
+        """Get the node features of neighboring nodes for each edge.
+        - node features of sender nodes (x[edge_index[0]])
+        - node features of receiver nodes (x[edge_index[1]])
         
-        A is the adjacency matrix of the graph.
+        edge_index denote the connectivity of the graph.
         """
         x, edge_index = x
         return x[edge_index[0]], x[edge_index[1]]
