+++ deeplay/components/gnn/pooling/__init__.py
@@ -1,2 +1,2 @@
-from .mincut import MinCutPooling
-from .graph_pool import *
+from .mincut import *
+from .globalpool import *
